fix filter id

+++ src/services/contacts.js
@@ -8,11 +8,12 @@ export const getAllContacts = async ({
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
   filter,
+  userId,
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
 
-  const contactsQuery = Contact.find();
+  const contactsQuery = Contact.find({ userId });
 
   if (filter.isFavourite) {
     contactsQuery.where('isFavourite').equals(filter.isFavourite);
@@ -48,7 +49,7 @@ export const getAllContacts = async ({
 };
 
 export const getContactById = async (contactId, userId) => {
-  return await Contact.findById(contactId, userId);
+  return await Contact.findByOne({ _id: contactId, userId });
 };
 
 export const createContact = async ({
@@ -57,6 +58,7 @@ export const createContact = async ({
   email,
   isFavourite = false,
   contactType,
+  userId, // Отримуємо userId
 }) => {
   const contact = new Contact({
     name,
@@ -64,13 +66,17 @@ export const createContact = async ({
     email,
     isFavourite,
     contactType,
-    userId: req.user._id,
+    userId, // Прив'язуємо контакт до користувача
   });
   return await contact.save(); // Збереження контакту у базу даних
 };
-export const updateContact = async (contactId, updateData) => {
-  return await Contact.findByIdAndUpdate(contactId, updateData, { new: true });
+export const updateContact = async (contactId, updateData, userId) => {
+  return await Contact.findByIdAndUpdate(
+    { _id: contactId, userId }, // Додано фільтр за userId
+    updateData,
+    { new: true },
+  );
 };
-export const deleteContact = async (id) => {
-  return await Contact.findByIdAndDelete(id);
+export const deleteContact = async (id, userId) => {
+  return await Contact.findByIdAndDelete({ _id: id, userId }); // Додано фільтр за userId
 };
